Load the blogs posts and list them on all a page
While the links to the blog posts will be the same as msanatan.com, we have to change the links to the blog pages.
Next.js crashes when I have 2 parameterised subpaths of /blog - maybe there's a config I haven't used.
This is fine, and likely better and more standard as well.